move old components ini DEPRECATED_components folder
