+++ src/DEPRECATED_components/JsccLogo/JsccLogo.js
