+++ src/DEPRECATED_components/JsccLogo/JsccLogo.story.js
