+++ src/DEPRECATED_components/JsccLogo/index.js
