+++ src/DEPRECATED_components/Location.js
