+++ src/DEPRECATED_components/Schedule.js
