+++ src/DEPRECATED_components/footer/footer.js
