+++ src/pages/values.js
@@ -1,8 +1,8 @@
 import React from "react"
 import { Link } from "gatsby"
 
-import Layout from "../components/layout"
-import SEO from "../components/seo"
+import Layout from "../DEPRECATED_components/layout"
+import SEO from "../DEPRECATED_components/seo"
 
 const Values = () => (
   <Layout>
